+++ app/src/main/java/com/jkdajac/englishlearning/MainActivity.kt
@@ -1,21 +1,27 @@
 package com.jkdajac.englishlearning
 
+import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.jkdajac.englishlearning.adapters.WordAdapter
 import com.jkdajac.englishlearning.database.worddb.AppDatabase
+import com.jkdajac.englishlearning.database.worddb.LearnedWords
 import com.jkdajac.englishlearning.database.worddb.Word
 import kotlinx.android.synthetic.main.activity_main.*
+import java.sql.RowId
 
 class MainActivity : AppCompatActivity(), WordAdapter.ViewHolder.ItemCallback{
     lateinit var adapter: WordAdapter
     lateinit var wordDatabase: AppDatabase
     lateinit var wordList: ArrayList<Word>
+    lateinit var learnedwordList: ArrayList<LearnedWords>
 
 
+    @SuppressLint("NotifyDataSetChanged")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
@@ -30,8 +36,10 @@ class MainActivity : AppCompatActivity(), WordAdapter.ViewHolder.ItemCallback{
         }
 
         getMyIntents()
-
+3
         wordList = ArrayList<Word>()
+        learnedwordList = ArrayList<LearnedWords>()
+        wordDatabase = AppDatabase.getDatabase(this)
         wordDatabase = AppDatabase.getDatabase(this)
         getData()
         adapter = WordAdapter(this, wordList, this)
@@ -40,6 +48,27 @@ class MainActivity : AppCompatActivity(), WordAdapter.ViewHolder.ItemCallback{
 
         wordDatabase = AppDatabase.getDatabase(this)
 
+        btGoToAllready.setOnClickListener {
+            val addDialog = AlertDialog.Builder(this)
+            addDialog
+                .setMessage("Вы действительно хотите перенести записи?")
+                .setPositiveButton("Ok") { dialog, _ ->
+                    getData()
+                    wordDatabase.wordDao().copy()
+                    wordDatabase.wordDao().deleteAll()
+                    finish();
+                    overridePendingTransition(0, 0);
+                    startActivity(intent);
+                    overridePendingTransition(0, 0);
+                    Toast.makeText(this, "Записи пересены!", Toast.LENGTH_SHORT).show()
+                    dialog.dismiss()
+                }
+                .setNegativeButton("Отмена") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                .create()
+                .show()
+        }
 
             btSave.setOnClickListener {
                 if (etEnglishWord.text.isNotEmpty() && etTranslateWord.text.isNotEmpty()) {
@@ -70,20 +99,33 @@ class MainActivity : AppCompatActivity(), WordAdapter.ViewHolder.ItemCallback{
         }
 
     override fun deleteItem(index: Int) {
-        val word = wordList.get(index)
-        wordDatabase.wordDao().copy()
-        wordDatabase.wordDao().deleteWord(word)
-        getData()
-        adapter.notifyDataSetChanged()
+
+        val addDialog = AlertDialog.Builder(this)
+        addDialog
+            .setMessage("Вы действительно хотите удалить запись?")
+            .setPositiveButton("Ok") { dialog, _ ->
+                val word = wordList[index]
+                wordDatabase.wordDao().deleteWord(word)
+                getData()
+                adapter.notifyDataSetChanged()
+                finish();
+                overridePendingTransition(0, 0);
+                startActivity(intent);
+                overridePendingTransition(0, 0);
+                Toast.makeText(this, "Запись удалена!", Toast.LENGTH_SHORT).show()
+                dialog.dismiss()
+            }
+            .setNegativeButton("Отмена") { dialog, _ ->
+                dialog.dismiss()
+            }
+            .create()
+            .show()
     }
 
-
     override fun openItem(index: Int) {
-
     }
 
     override fun closeItem(index: Int) {
-        
     }
 
     fun getMyIntents(){
